pipeline: move pipeline CLI wire-ups into a "cli" module here
