--- docs/api/toasty.cli.cascade_getparser.rst
@@ -1,6 +1,0 @@
-cascade_getparser
-=================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: cascade_getparser
