--- docs/api/toasty.cli.cascade_impl.rst
@@ -1,6 +1,0 @@
-cascade_impl
-============
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: cascade_impl
