--- docs/api/toasty.cli.make_thumbnail_getparser.rst
@@ -1,6 +1,0 @@
-make_thumbnail_getparser
-========================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: make_thumbnail_getparser
