--- docs/api/toasty.cli.make_thumbnail_impl.rst
@@ -1,6 +1,0 @@
-make_thumbnail_impl
-===================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: make_thumbnail_impl
