--- docs/api/toasty.cli.multi_tan_make_data_tiles_getparser.rst
@@ -1,6 +1,0 @@
-multi_tan_make_data_tiles_getparser
-===================================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: multi_tan_make_data_tiles_getparser
