--- docs/api/toasty.cli.multi_tan_make_data_tiles_impl.rst
@@ -1,6 +1,0 @@
-multi_tan_make_data_tiles_impl
-==============================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: multi_tan_make_data_tiles_impl
