--- docs/api/toasty.cli.tile_allsky_getparser.rst
@@ -1,6 +1,0 @@
-tile_allsky_getparser
-=====================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_allsky_getparser
