--- docs/api/toasty.cli.tile_allsky_impl.rst
@@ -1,6 +1,0 @@
-tile_allsky_impl
-================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_allsky_impl
