--- docs/api/toasty.cli.tile_healpix_getparser.rst
@@ -1,6 +1,0 @@
-tile_healpix_getparser
-======================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_healpix_getparser
