--- docs/api/toasty.cli.tile_healpix_impl.rst
@@ -1,6 +1,0 @@
-tile_healpix_impl
-=================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_healpix_impl
