--- docs/api/toasty.cli.tile_study_getparser.rst
@@ -1,6 +1,0 @@
-tile_study_getparser
-====================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_study_getparser
