--- docs/api/toasty.cli.tile_study_impl.rst
@@ -1,6 +1,0 @@
-tile_study_impl
-===============
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_study_impl
