--- docs/api/toasty.cli.tile_wwtl_getparser.rst
@@ -1,6 +1,0 @@
-tile_wwtl_getparser
-===================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_wwtl_getparser
