--- docs/api/toasty.cli.tile_wwtl_impl.rst
@@ -1,6 +1,0 @@
-tile_wwtl_impl
-==============
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: tile_wwtl_impl
