--- docs/api/toasty.cli.transform_getparser.rst
@@ -1,6 +1,0 @@
-transform_getparser
-===================
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: transform_getparser
