--- docs/api/toasty.cli.transform_impl.rst
@@ -1,6 +1,0 @@
-transform_impl
-==============
-
-.. currentmodule:: toasty.cli
-
-.. autofunction:: transform_impl
